base, orm, view: Implement ORM data source layer
I'm not too happy with the result (especially the amount of boilerplate
I had to put in tables.ts), but it works fine. It was a pretty steep
learning curve though, because Sequelize wants to manage the database by
itself, and it's a tough beast to teach giving you the reins instead of
automatically guessing things like pluralized table names.